% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-name.R
\name{extract_name}
\alias{extract_name}
\title{Extract patient name from a character string}
\usage{
extract_name(input)
}
\arguments{
\item{input}{A character string containing the patient name after the labno}
}
\value{
The patient name as a string
}
\description{
Extract patient name from a character string
}
\examples{
extract_name("WS123456_12345678_AnnaKarenina")
}
