% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename-to-df.R
\name{filename_to_df}
\alias{filename_to_df}
\title{Convert a filename into a dataframe of identifiers}
\usage{
filename_to_df(filename)
}
\arguments{
\item{filename}{A filename with identifiers in specified format}
}
\value{
A dataframe of sample identifiers
}
\description{
Convert a filename into a dataframe of identifiers
}
\examples{
filename_to_df("Annotated_WS123456_12345678a_PierreBEZUKHOV.xlsx")
}
