% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-ids.R
\name{mutate_ids}
\alias{mutate_ids}
\title{Add sample identifiers from filename as additional data columns}
\usage{
mutate_ids(df, id_col = filename, regex = regex_filename()$standard_regex)
}
\arguments{
\item{df}{The dataframe to add identifiers to}

\item{id_col}{The column in the dataframe which includes the filename}

\item{regex}{The regular expression to use for identifier matching}
}
\value{
The original dataframe with additional columns of identifiers
from the filename column
}
\description{
Add sample identifiers from filename as additional data columns
}
