% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex-filename.R
\name{regex_filename}
\alias{regex_filename}
\title{Regular expressions for filenames}
\usage{
regex_filename()
}
\value{
A named list of regular expressions for different filename formats
}
\description{
Regular expressions for filenames
}
\examples{
regex_filename()$filename_with_names
}
