% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-ids.R
\name{mutate_ids}
\alias{mutate_ids}
\title{Add sample identifiers from filename as additional data columns}
\usage{
mutate_ids(df, filename)
}
\arguments{
\item{df}{The dataframe to add identifiers to}

\item{filename}{The filename of the dataframe}
}
\value{
The original dataframe with additional columns of identifiers
from the filename
}
\description{
Add sample identifiers from filename as additional data columns
}
\examples{

data_df <- data.frame(
"col1" = c(1, 2),
"col2" = c("a", "b"))

mutate_ids(df = data_df,
filename = "Annotated_WS123456_12345678a_PierreBEZUKHOV.xlsx")

}
